Use floats instead of doubles for our update time values.
They're quite a bit faster (on current iPhone hardware at least)

+++ Classes/BTAlphaTask.h
@@ -5,9 +5,9 @@
 #import "BTInterpolationTask.h"
 
 @interface BTAlphaTask : BTInterpolationTask
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha;
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha withInterpolator:(BTInterpolator)interp;
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha onDisplay:(SPDisplayObject*)display;
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha;
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha withInterpolator:(BTInterpolator)interp;
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha onDisplay:(SPDisplayObject*)display;
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha
   withInterpolator:(BTInterpolator)interp onDisplay:(SPDisplayObject*)display;
 @end
