+++ Classes/BTAlphaTask.m
@@ -8,19 +8,19 @@
     float _deltaAlpha;
 }
 
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha {
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha {
     return [self initOverTime:seconds toAlpha:alpha withInterpolator:BTLinearInterpolator];
 }
 
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha withInterpolator:(BTInterpolator)interp {
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha withInterpolator:(BTInterpolator)interp {
     return [self initOverTime:seconds toAlpha:alpha withInterpolator:interp onDisplay:nil];
 }
 
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha onDisplay:(SPDisplayObject*)display {
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha onDisplay:(SPDisplayObject*)display {
     return [self initOverTime:seconds toAlpha:alpha withInterpolator:BTLinearInterpolator onDisplay:display];
 }
 
-- (id)initOverTime:(double)seconds toAlpha:(float)alpha
+- (id)initOverTime:(float)seconds toAlpha:(float)alpha
   withInterpolator:(BTInterpolator)interp onDisplay:(SPDisplayObject*)display {
     if (!(self = [super initOverTime:seconds withInterpolator:interp onDisplay:display])) return nil;
     [self.attached connectUnit:^{
