+++ Classes/BTMode+Protected.h
@@ -2,5 +2,5 @@
 //  Betwixt - Copyright 2011 Three Rings Design
 
 @interface BTMode (protected)
-- (void)update:(double)dt;
+- (void)update:(float)dt;
 @end
